+++ src/controller/soal.js
@@ -1,13 +1,17 @@
 import soalModel from "../models/soal.js";
 
 const addSoal = async (req, res) => {
-    const { body } = req;
+    const { body, user } = req;
     const { params } = req;
     const { id_bank_soal } = params;
 
     try {
-        const result = await soalModel.addSoal(id_bank_soal, body);
-        res.status(201).json({ success: true, message: 'Soal baru telah ditambahkan', result });
+        if (user && user.role === 'Kontributor') {
+            const result = await soalModel.addSoal(id_bank_soal, body);
+            res.status(201).json({ success: true, message: 'Soal baru telah ditambahkan', result });
+        } else {
+            res.status(403).json({ success: false, message: 'Unauthorized' });
+        }
     } catch (error) {
         console.error(error);
         res.status(500).json({ success: false, message: 'Gagal menambahkan soal', error: error.message });
@@ -15,13 +19,17 @@ const addSoal = async (req, res) => {
 };
 
 const updateSoal = async (req, res) => {
-    const { body, params } = req;
+    const { body, params, user } = req;
     const { id_soal } = params;
 
     try {
-        // Pastikan model Anda dapat menangani data dari req.body tanpa perlu mengonversi JSON
-        const result = await soalModel.updateSoal(id_soal, body);
-        res.status(201).json({ success: true, message: 'Soal telah diperbarui', result });
+        if (user && user.role === 'Kontributor') {
+            // Pastikan model Anda dapat menangani data dari req.body tanpa perlu mengonversi JSON
+            const result = await soalModel.updateSoal(id_soal, body);
+            res.status(201).json({ success: true, message: 'Soal telah diperbarui', result });
+        } else {
+            res.status(403).json({ success: false, message: 'Unauthorized' });
+        }
     } catch (error) {
         console.error(error);
         res.status(500).json({ success: false, message: 'Gagal memperbarui soal', error: error.message });
@@ -43,13 +51,17 @@ const getAllSoal = async (req, res) => {
 
 const deleteSoal = async (req, res) => {
     const { id_soal } = req.params;
+    const { user } = req;
 
     try {
-        // Panggil fungsi deleteSoal dengan id_soal yang diberikan
-        await soalModel.deleteSoal(id_soal);
-
-        // Berikan respons sukses jika tidak ada kesalahan
-        res.status(200).json({ message: 'Soal berhasil dihapus.' });
+        if (user && user.role === 'Kontributor') {
+            // Panggil fungsi deleteSoal dengan id_soal yang diberikan
+            await soalModel.deleteSoal(id_soal);
+            // Berikan respons sukses jika tidak ada kesalahan
+            res.status(200).json({ message: 'Soal berhasil dihapus.' });
+        } else {
+            res.status(403).json({ success: false, message: 'Unauthorized' });
+        }
     } catch (error) {
         // Tangani kesalahan dan kirim respons dengan status error
         console.error(error);
